+++ packages/mobile_sidebar/lib/common/index.dart
@@ -1,2 +1,3 @@
-export 'action_button.dart';
+export 'action_item.dart';
+export 'menu_item.dart';
 export 'raised_card.dart';