+++ packages/mobile_popup/CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.1.0
+
+* Breaking change changing builder to child on `1.0.2`
+
+# 1.0.2
+
+* Adding Fullscreen Option `showFullscreen`
+* Updated Example
+
+# 1.0.1
+
+* Fixing root navigator
+
+# 1.0.0
+
+* Added support for `mobilePush` which on iOS and Android when set to `true` will push to a screen instead of showing a pop up.
+
 ## 0.1.0
 
 * Added Desktop Targets