+++ packages/mobile_sidebar/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.5.0 - 05.15.2019
+
+* Adding Option to Persist Index
+* Updated Example
+* Removed Staggered Grid View
+* Updating UI of Grid Tiles
+* Added `mobileBottomNavigation` to show BottomNavigationBar on mobile
+
 ## 1.0.0
 
 * Creating new widget